quant-trading/frontend/requirements.txt
streamlit==1.29.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.3
plotly==5.18.0
pillow==10.0.1
pydantic==2.5.2
jinja2==3.1.3
watchdog==3.0.0
